Adds def item to place serializer
Added a definition for item in the place controller to that when it is
serialized you see the id of the item with the place. This is so in
Ember it will pull the items associated with place. So far in a `show`
action in curl it had show the item id with the place.